Adding tests

+++ tools-init.sh
@@ -10,11 +10,18 @@ export AWS_REGION=$(curl -s http://169.254.169.254/latest/meta-data/placement/av
 if [ -z "$AWS_REGION" ]
 then
     # metadata might err, this is a safeguard
+    echo "Error: AWS region not found, exiting"
     exit 0
+else
+    echo "Deploying into AWS region: ${AWS_REGION}"
 fi
 
 # Export Default Env Variables ------------------
-echo "export AWS_REGION=${AWS_REGION}" >> ~/.bash_profile
+if ! grep -q 'export AWS_REGION' ~/.bash_profile; then
+    echo "export AWS_REGION=${AWS_REGION}" >> ~/.bash_profile
+fi
+
+
 aws configure set default.region ${AWS_REGION}
 aws configure get default.region
 
@@ -22,39 +29,77 @@ export NVM_VER=$(curl --silent "https://github.com/nvm-sh/nvm/releases/latest" |
 export SWB_VER=$(curl --silent "https://github.com/awslabs/service-workbench-on-aws/releases/latest" | sed 's#.*tag/\(.*\)\".*#\1#') #v.3.1.0
 export PACKER_VER=1.7.2
 
-# Clone SWB and install dependencies ------------
-echo "Cloning SWB Repo ..."
-cd ~/environment
-#git clone --depth 1 --branch $SWB_VER https://github.com/awslabs/service-workbench-on-aws.git >/dev/null 2>&1
-git clone https://github.com/awslabs/service-workbench-on-aws.git >/dev/null 2>&1
+# Ensure SWB code exists.  Assume it's the latest version. ------------
+SWB_DIR=~/environment/service-workbench-on-aws
+if [ -d $SWB_DIR ]; then
+    cd $SWB_DIR
+    CURRENT_VER=$(git describe --tags --abbrev=0)
+    echo "SWB code ${CURRENT_VER} already installed"
+else
+    echo "Cloning SWB Repo ${SWB_VER} from GitHub into ~/environment"
+    cd ~/environment
+    git clone https://github.com/awslabs/service-workbench-on-aws.git >/dev/null 2>&1
+fi
 cd $cwd
-echo "Installing dependencies ..."
-sudo yum install golang jq -y -q -e 0 >/dev/null 2>&1
+
+DEPENDENCIES=(golang jq)
+# echo "Installing dependencies ${DEPENDENCIES} ..."
+for dependency in ${DEPENDENCIES[@]}; do
+    if $(! yum list installed $dependency > /dev/null 2>&1); then
+	echo "Installing dependency: $dependency"
+	sudo yum install $dependency -y -q -e 0 >/dev/null 2>&1
+    else
+	echo "Dependency $dependency exists"
+    fi
+done
+
 echo "Enabling utilities scripts ..."
 chmod +x cloud9-resize.sh
 chmod +x hosting-account/create-host-account.sh
-echo "Resizing AWS Cloud9 Volume ..."
-./cloud9-resize.sh #50GB by default
+
+DISKSIZE=$(df -m . | tail -1 | awk '{print $2}')
+if (( DISKSIZE > 40000 )); then
+    echo "Installation volume has adequate size: ${DISKSIZE} MB"
+else
+    echo "Resizing AWS Cloud9 Volume to 50 GB ..."    
+    ./cloud9-resize.sh #50GB by default
+fi
 
 # NVM & Node Versions ---------------------------
-echo "Installing nvm ..."
-rm -rf ~/.nvm
-export NVM_DIR=
-curl --silent -o- "https://raw.githubusercontent.com/nvm-sh/nvm/$NVM_VER/install.sh" | bash
-source ~/.nvm/nvm.sh 
-nvm install --lts 
+# LTS_VER=$(nvm version-remote --lts)
+source ~/.nvm/nvm.sh
+if ! nvm --version > /dev/null 2>&1; then
+    echo "Installing nvm ${NVM_VER} ..."
+    rm -rf ~/.nvm
+    export NVM_DIR=
+    curl --silent -o- "https://raw.githubusercontent.com/nvm-sh/nvm/$NVM_VER/install.sh" | bash
+    source ~/.nvm/nvm.sh
+else
+    echo "nvm version $(nvm -v) is installed"
+fi
+exit
+
+nvm install --lts
+nvm use --lts
 nvm alias default stable
+node --version
+echo "Exiting, check node version"
+exit
 
 # npm packages ----------------------------------
+NPM_PACKAGES=(serverless pnpm hygen yarn docusaurus)
 echo "Installing framework and libs ..."
-npm install -g serverless pnpm hygen yarn docusaurus >/dev/null 2>&1
+npm install -g $NPM_PACKAGES >/dev/null 2>&1
 
 # packer ----------------------------------------
-echo "Installing packer ..."
+echo "Installing packer ${PACKER_VER} into /usr/local/bin/ ..."
 wget -q "https://releases.hashicorp.com/packer/$PACKER_VER/packer_${PACKER_VER}_linux_amd64.zip" -O packer_${PACKER_VER}_linux_amd64.zip
 unzip "packer_${PACKER_VER}_linux_amd64.zip" >/dev/null 2>&1
 sudo mv packer /usr/local/bin/ >/dev/null 2>&1
 rm -f "packer_${PACKER_VER}_linux_amd64.zip" >/dev/null 2>&1
+
+echo "Exiting"
+exit
 
 # finishing up ----------------------------------
 echo "Finishing up ..."
